convert to long

+++ features/merchant/seller_menu_common/src/main/java/com/tokopedia/seller/menu/common/domain/entity/UserShopInfoResponse.kt
@@ -71,7 +71,7 @@ data class UserShopInfoResponse(
                     val startTime: String = "",
                     @Expose
                     @SerializedName("value")
-                    val value: Int = 0
+                    val value: Long = 0
             )
         }
     }
